requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.48.0
django-appconf==1.0.5
django-crispy-forms==1.13.0
django-dotenv==1.4.2
django-htmx==1.8.0
django-rest-framework==0.1.0
django-select2==7.10.0
django-storages==1.12.3
django-userforeignkey==0.4.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
oauthlib==3.2.0
openpyxl==3.0.9
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
